Move combineLatestAll into CombineLatest
The extension name is a publicly visible artifact and changing it will
be breaking. It probably makes sense to combine these now.
